concierge_agent/supervisor_agent/requirements.txt
bedrock-agentcore>=0.1.2
bedrock-agentcore-starter-toolkit>=0.1.6
boto3>=1.40.8
pydantic>=2.0.0
pydantic-core>=2.33.2
pytz>=2025.2
rapidfuzz>=3.13.0
requests-aws4auth>=1.3.1
strands-agents>=1.4.0
strands-agents-tools>=0.2.3
uv
mcp
requests
gql>=3.4.0
requests>=2.28.0
tavily-python
langgraph>=0.2.0
langchain-aws>=0.2.0
langchain-core>=0.3.0
# playwright>=1.40.0